move supabase config
