+++ backend/internal/supabase/.gitignore
@@ -1,0 +1,4 @@
+# Supabase
+.branches
+.temp
+.env
