+++ backend/internal/supabase/seed.sql
@@ -1,0 +1,7 @@
+INSERT INTO users (user_id, first_name, last_name, email, phone, profile_picture)
+VALUES
+  ('user1', 'John', 'Smith', 'john.smith@example.com', '123-456-7890', NULL),
+  ('user2', 'Jane', 'Doe', 'jane.doe@example.com', '987-654-3210', NULL),
+  ('user3', 'Bob', 'Johnson', 'bob.johnson@example.com', NULL, NULL, NULL),
+  ('user4', 'Emily', 'Garcia', 'emily.garcia@example.com', '555-1212', NULL)
+;
